+++ src/main/java/com/nekohit/neo/helper/Utils.java
@@ -1,15 +1,9 @@
 package com.nekohit.neo.helper;
 
-import io.neow3j.crypto.ECKeyPair;
 import io.neow3j.devpack.ByteString;
 import io.neow3j.devpack.annotations.Instruction;
 import io.neow3j.script.OpCode;
-import io.neow3j.transaction.Transaction;
 import io.neow3j.types.StackItemType;
-import io.neow3j.wallet.Account;
-
-import java.math.BigInteger;
-import java.util.Scanner;
 
 public class Utils {
     @Instruction(opcode = OpCode.CONVERT, operand = StackItemType.BYTE_STRING_CODE)
@@ -36,32 +30,5 @@ public class Utils {
         if (!condition) {
             throw new Exception(message);
         }
-    }
-
-    public static Account createAccountFromPrivateKey(String privateKey) {
-        return new Account(ECKeyPair.create(
-                new BigInteger(privateKey, 16)
-        ));
-    }
-
-    /**
-     * Get gas fee from given test
-     *
-     * @param tx the transaction
-     * @return the gas fee in decimal representation
-     */
-    public static double getGasFeeFromTx(Transaction tx) {
-        long fraction = tx.getSystemFee() + tx.getNetworkFee();
-        return fraction / Math.pow(10, 8);
-    }
-
-    public static Account readAccountWIF(Scanner scanner) {
-        System.out.println("Paste account WIF:");
-        String walletWIF = scanner.nextLine();
-        // flush WIF out of screen
-        for (int i = 0; i < 2000; i++) {
-            System.out.println();
-        }
-        return Account.fromWIF(walletWIF);
     }
 }
