+++ src/test/java/com/nekohit/neo/contract/WCACreateTest.java
@@ -129,8 +129,7 @@ public class WCACreateTest extends ContractTestFramework {
                         },
                         new Signer[]{
                                 AccountSigner.calledByEntry(this.testWallet.getDefaultAccount())
-                        },
-                        this.testWallet
+                        }
                 )
         );
         assertTrue(
